Start of test cleanup
